+++ R/generics.R
@@ -214,6 +214,7 @@ dplyr_reconstruct.rowwise_df <- function(data, template) {
 
 dplyr_col_select <- function(.data, loc, names = NULL) {
   loc <- vec_as_location(loc, n = ncol(.data), names = names(.data))
+
   out <- .data[loc]
   if (!inherits(out, "data.frame")) {
     abort(c(
@@ -239,8 +240,16 @@ dplyr_col_select <- function(.data, loc, names = NULL) {
       )
     ))
   }
+
+  # Patch base data frames to restore extra attributes that `[.data.frame` drops.
+  # We require `[` methods to keep extra attributes for all data frame subclasses.
+  if (identical(class(.data), "data.frame")) {
+    out <- dplyr_reconstruct(out, .data)
+  }
+
   if (!is.null(names)) {
     names(out) <- names
   }
+
   out
 }